fix(ibkr): place_market_order and place_limit_order units type int→float
BaseBroker interface declares units as float (for crypto fractional
quantities), but IBKRClient had int type annotations on place_market_order
and place_limit_order. This was missed in the earlier fix 7efd311 which
only corrected position retrieval truncation. Python's dynamic typing
masked the issue at runtime, but the annotation was wrong and could
break static type checkers or mislead callers.

+++ backend/broker/ibkr_client.py
@@ -576,7 +576,7 @@ class IBKRClient(BaseBroker):
     async def place_market_order(
         self,
         instrument: str,
-        units: int,
+        units: float,
         stop_loss: Optional[float] = None,
         take_profit: Optional[float] = None,
     ) -> OrderResult:
@@ -660,7 +660,7 @@ class IBKRClient(BaseBroker):
     async def place_limit_order(
         self,
         instrument: str,
-        units: int,
+        units: float,
         price: float,
         stop_loss: Optional[float] = None,
         take_profit: Optional[float] = None,
